Move tuple tests back
